+++ src/tileEntity/tileEntity.java
@@ -8,9 +8,9 @@ public class TileEntity implements Serializable{
 	public short textureID;
 	public byte tileWidth, tileHeight;
 	public byte xOffset, yOffset;
-	public byte x, y;
+	public int x, y;
 
-	public TileEntity(byte x, byte y, byte tileWidth, byte tileHeight, short tileEntityID){
+	public TileEntity(int x, int y, byte tileWidth, byte tileHeight, short tileEntityID){
 
 		this.tileHeight = tileHeight;
 		this.tileWidth = tileWidth;
